// src/view/after_login/admin/daftar_vendor/model/ModelAddVendor.ts
export interface ModelAddVendor {
    no_vendor: string;
    nama_vendor: string;
    pemilik_vendor: string;
    alamat: string;
    telpon: string;
}


export type TypeModelVendor = {
    no_vendor: string;
    nama_vendor: string;
    pemilik_vendor: string;
    alamat: string;
    telpon: string;
}